Make the LayerList test_basic() more robust.
Instead of taking the first layer of the OpenLayers.Map and compair it to the
first/last record of the GXM.Map's layerstore, we first find the maximum and
minimum z-indeces of the OPenLayers.Map and use this to compair with the first
or last (depending on sort direction) records z-index.

This way the test passes with current HEAD and with commit
b8f710f151af165e382eed6c0c6ba27c39f7aa0d applied.

See also #31; if this and the commits from #31 get merged, all tests pass.